Havasam nabood kolli chiz doros shode yadam nemiad kodoma bood :|

+++ app/Http/Controllers/API/V01/Thread/AnswerController.php
@@ -4,9 +4,16 @@ namespace App\Http\Controllers\API\V01\Thread;
 
 use App\Models\Answer;
 use App\Models\Thread;
+use App\Models\Subscribe;
+use App\Models\User;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use App\Repositories\AnswerRepository;
+use App\Notifications\NewReplySubmitted;
+use App\Repositories\SubscribeRepository;
+use App\Repositories\UserRepository;
+use Illuminate\Support\Facades\Notification;
 use Symfony\Component\HttpFoundation\Response;
 
 class AnswerController extends Controller
@@ -32,7 +39,15 @@ class AnswerController extends Controller
         ]);
 
         resolve(AnswerRepository::class)->store($request);
+        // List users_id which subscribed to a thread id
+        $notifiable_users_id = resolve(SubscribeRepository::class)->getNotifiableUser($request->thread_id);
 
+        // Get user instance from id 
+        $notifiable_users = resolve(UserRepository::class)->find($notifiable_users_id);
+        
+        // Send NewReplySubmitted Notification To Subscribed Users
+        Notification::send($notifiable_users, new NewReplySubmitted(Thread::find($request->thread_id)));
+        
         return response()->json([
             'message' => 'answers created successfully',
         ], Response::HTTP_CREATED);
@@ -44,22 +59,36 @@ class AnswerController extends Controller
         $request->validate([
             'content' => 'required',
         ]);
+        
+        if(Gate::forUser(auth()->user())->allows('user-answer', $answer)){
 
-        resolve(AnswerRepository::class)->update($request, $answer);
+            resolve(AnswerRepository::class)->update($request, $answer);
 
+            return response()->json([
+                'message' => 'answer updated successfully',
+            ],Response::HTTP_OK);
+        }
 
         return response()->json([
-            'message' => 'answer updated successfully',
-        ],Response::HTTP_OK);
+            'message' => 'access denied',
+        ],Response::HTTP_FORBIDDEN);
     }
 
 
     public function destroy(Answer $answer)
     {
-        resolve(AnswerRepository::class)->destroy($answer);
-        
+        if(Gate::forUser(auth()->user())->allows('user-answer', $answer)){
+
+            resolve(AnswerRepository::class)->destroy($answer);
+
+            return response()->json([
+                'message' => 'answer deleted successfully',
+            ],Response::HTTP_OK);
+    
+        }
         return response()->json([
-            'message' => 'answer deleted successfully',
-        ],Response::HTTP_OK);
+            'message' => 'access denied',
+        ],Response::HTTP_FORBIDDEN);
+
     }
 }
